adding testGUI project -Dennis
This screen shows what the employee will see most of the time.  The
keypad on the screen will be used to enter or remove item numbers.  The
frame on the left is used to view the current transaction.  The buttons
on the right will serve as shortcuts to items if they would prefer not
to enter the numbers in.